requirements.txt
# 多智能体医疗协作系统 - Python依赖列表
# 更新时间: 2024年12月30日
# 项目状态: 生产就绪

# 强化学习框架
gymnasium>=0.29.0
stable-baselines3>=2.0.0
pettingzoo>=1.24.0

# 数据处理和科学计算
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# 深度学习
torch>=2.0.0

# 可视化
matplotlib>=3.7.0
seaborn>=0.12.0

# Web演示框架
flask>=2.3.0
flask-socketio>=5.3.0
python-socketio>=5.8.0

# 配置管理
pyyaml>=6.0.0

# 进度条和用户界面
tqdm>=4.65.0

# 环境变量管理 (可选)
python-dotenv>=1.0.0

# 注释：已移除的依赖
# supersuit>=3.9.0         # 多智能体包装器 - 本项目未使用
# scipy>=1.10.0           # 科学计算 - 本项目未直接使用
# plotly>=5.15.0          # 交互式图表 - 使用Chart.js替代
# jupyter>=1.0.0          # Jupyter笔记本 - 生产环境不需要
# wandb>=0.15.0           # 权重和偏置跟踪 - 本项目未使用
# tensorboard>=2.10.0     # TensorBoard - Stable-Baselines3自带 